Refactor shell completion into dedicated module runtime
