+++ src/Repl.Core/ShellCompletion/IShellCompletionAdapter.cs
@@ -1,0 +1,17 @@
+namespace Repl.ShellCompletion;
+
+internal interface IShellCompletionAdapter
+{
+	ShellKind Kind { get; }
+
+	string Token { get; }
+
+	string ResolveProfilePath(
+		ShellCompletionOptions options,
+		bool parentLooksLikeWindowsPowerShell,
+		string userHomePath);
+
+	string BuildManagedBlock(string commandName, string appId);
+
+	string BuildReloadHint();
+}
